requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyjwt==2.8.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.0
